# Optimize tests and fix flaky tests to reduce race conditions