### Summary
Some people were complaining that the tests where taking too long to run, so this goes out for them :P

#### Local machine (no coverage) [`nox -s pytest -- --skip-coverage`]
```
Before: 11s 264ms
After: 996ms
```
#### CI (https://github.com/hikari-py/hikari/runs/5074464193 vs https://github.com/hikari-py/hikari/runs/5067388127) [ubuntu-latest 3.8 - "Run tests" step]
```
Before: 2min 14s
After: 1min 49s
```


#
This pr also removes `TimedCacheMap`, as while optimizing the tests for it I noticed that it wasnt being used anywhere. This removal can be moved to a separate pr if needed. Since its considered internal detail, I dont think there is a need for deprecation or breaking changelog

### Checklist
<!-- Make sure to tick all the following boxes by putting an `x` in between (like this `[x]`) -->
- [x] I have run `nox` and all the pipelines have passed.
- [x] I have made unittests according to the code I have added/modified/deleted.

### Related issues
Closes #1010